--- a/modules/apps/asset/asset-display-page-api/src/main/java/com/liferay/asset/display/page/service/persistence/AssetDisplayPageEntryUtil.java
+++ b/modules/apps/asset/asset-display-page-api/src/main/java/com/liferay/asset/display/page/service/persistence/AssetDisplayPageEntryUtil.java
@@ -904,6 +904,196 @@ public class AssetDisplayPageEntryUtil {
 	}
 
 	/**
+	 * Returns all the asset display page entries where groupId = &#63; and classNameId = &#63;.
+	 *
+	 * @param groupId the group ID
+	 * @param classNameId the class name ID
+	 * @return the matching asset display page entries
+	 */
+	public static List<AssetDisplayPageEntry> findByG_CN(
+		long groupId, long classNameId) {
+
+		return getPersistence().findByG_CN(groupId, classNameId);
+	}
+
+	/**
+	 * Returns a range of all the asset display page entries where groupId = &#63; and classNameId = &#63;.
+	 *
+	 * <p>
+	 * Useful when paginating results. Returns a maximum of <code>end - start</code> instances. <code>start</code> and <code>end</code> are not primary keys, they are indexes in the result set. Thus, <code>0</code> refers to the first result in the set. Setting both <code>start</code> and <code>end</code> to <code>QueryUtil#ALL_POS</code> will return the full result set. If <code>orderByComparator</code> is specified, then the query will include the given ORDER BY logic. If <code>orderByComparator</code> is absent, then the query will include the default ORDER BY logic from <code>AssetDisplayPageEntryModelImpl</code>.
+	 * </p>
+	 *
+	 * @param groupId the group ID
+	 * @param classNameId the class name ID
+	 * @param start the lower bound of the range of asset display page entries
+	 * @param end the upper bound of the range of asset display page entries (not inclusive)
+	 * @return the range of matching asset display page entries
+	 */
+	public static List<AssetDisplayPageEntry> findByG_CN(
+		long groupId, long classNameId, int start, int end) {
+
+		return getPersistence().findByG_CN(groupId, classNameId, start, end);
+	}
+
+	/**
+	 * Returns an ordered range of all the asset display page entries where groupId = &#63; and classNameId = &#63;.
+	 *
+	 * <p>
+	 * Useful when paginating results. Returns a maximum of <code>end - start</code> instances. <code>start</code> and <code>end</code> are not primary keys, they are indexes in the result set. Thus, <code>0</code> refers to the first result in the set. Setting both <code>start</code> and <code>end</code> to <code>QueryUtil#ALL_POS</code> will return the full result set. If <code>orderByComparator</code> is specified, then the query will include the given ORDER BY logic. If <code>orderByComparator</code> is absent, then the query will include the default ORDER BY logic from <code>AssetDisplayPageEntryModelImpl</code>.
+	 * </p>
+	 *
+	 * @param groupId the group ID
+	 * @param classNameId the class name ID
+	 * @param start the lower bound of the range of asset display page entries
+	 * @param end the upper bound of the range of asset display page entries (not inclusive)
+	 * @param orderByComparator the comparator to order the results by (optionally <code>null</code>)
+	 * @return the ordered range of matching asset display page entries
+	 */
+	public static List<AssetDisplayPageEntry> findByG_CN(
+		long groupId, long classNameId, int start, int end,
+		OrderByComparator<AssetDisplayPageEntry> orderByComparator) {
+
+		return getPersistence().findByG_CN(
+			groupId, classNameId, start, end, orderByComparator);
+	}
+
+	/**
+	 * Returns an ordered range of all the asset display page entries where groupId = &#63; and classNameId = &#63;.
+	 *
+	 * <p>
+	 * Useful when paginating results. Returns a maximum of <code>end - start</code> instances. <code>start</code> and <code>end</code> are not primary keys, they are indexes in the result set. Thus, <code>0</code> refers to the first result in the set. Setting both <code>start</code> and <code>end</code> to <code>QueryUtil#ALL_POS</code> will return the full result set. If <code>orderByComparator</code> is specified, then the query will include the given ORDER BY logic. If <code>orderByComparator</code> is absent, then the query will include the default ORDER BY logic from <code>AssetDisplayPageEntryModelImpl</code>.
+	 * </p>
+	 *
+	 * @param groupId the group ID
+	 * @param classNameId the class name ID
+	 * @param start the lower bound of the range of asset display page entries
+	 * @param end the upper bound of the range of asset display page entries (not inclusive)
+	 * @param orderByComparator the comparator to order the results by (optionally <code>null</code>)
+	 * @param useFinderCache whether to use the finder cache
+	 * @return the ordered range of matching asset display page entries
+	 */
+	public static List<AssetDisplayPageEntry> findByG_CN(
+		long groupId, long classNameId, int start, int end,
+		OrderByComparator<AssetDisplayPageEntry> orderByComparator,
+		boolean useFinderCache) {
+
+		return getPersistence().findByG_CN(
+			groupId, classNameId, start, end, orderByComparator,
+			useFinderCache);
+	}
+
+	/**
+	 * Returns the first asset display page entry in the ordered set where groupId = &#63; and classNameId = &#63;.
+	 *
+	 * @param groupId the group ID
+	 * @param classNameId the class name ID
+	 * @param orderByComparator the comparator to order the set by (optionally <code>null</code>)
+	 * @return the first matching asset display page entry
+	 * @throws NoSuchDisplayPageEntryException if a matching asset display page entry could not be found
+	 */
+	public static AssetDisplayPageEntry findByG_CN_First(
+			long groupId, long classNameId,
+			OrderByComparator<AssetDisplayPageEntry> orderByComparator)
+		throws com.liferay.asset.display.page.exception.
+			NoSuchDisplayPageEntryException {
+
+		return getPersistence().findByG_CN_First(
+			groupId, classNameId, orderByComparator);
+	}
+
+	/**
+	 * Returns the first asset display page entry in the ordered set where groupId = &#63; and classNameId = &#63;.
+	 *
+	 * @param groupId the group ID
+	 * @param classNameId the class name ID
+	 * @param orderByComparator the comparator to order the set by (optionally <code>null</code>)
+	 * @return the first matching asset display page entry, or <code>null</code> if a matching asset display page entry could not be found
+	 */
+	public static AssetDisplayPageEntry fetchByG_CN_First(
+		long groupId, long classNameId,
+		OrderByComparator<AssetDisplayPageEntry> orderByComparator) {
+
+		return getPersistence().fetchByG_CN_First(
+			groupId, classNameId, orderByComparator);
+	}
+
+	/**
+	 * Returns the last asset display page entry in the ordered set where groupId = &#63; and classNameId = &#63;.
+	 *
+	 * @param groupId the group ID
+	 * @param classNameId the class name ID
+	 * @param orderByComparator the comparator to order the set by (optionally <code>null</code>)
+	 * @return the last matching asset display page entry
+	 * @throws NoSuchDisplayPageEntryException if a matching asset display page entry could not be found
+	 */
+	public static AssetDisplayPageEntry findByG_CN_Last(
+			long groupId, long classNameId,
+			OrderByComparator<AssetDisplayPageEntry> orderByComparator)
+		throws com.liferay.asset.display.page.exception.
+			NoSuchDisplayPageEntryException {
+
+		return getPersistence().findByG_CN_Last(
+			groupId, classNameId, orderByComparator);
+	}
+
+	/**
+	 * Returns the last asset display page entry in the ordered set where groupId = &#63; and classNameId = &#63;.
+	 *
+	 * @param groupId the group ID
+	 * @param classNameId the class name ID
+	 * @param orderByComparator the comparator to order the set by (optionally <code>null</code>)
+	 * @return the last matching asset display page entry, or <code>null</code> if a matching asset display page entry could not be found
+	 */
+	public static AssetDisplayPageEntry fetchByG_CN_Last(
+		long groupId, long classNameId,
+		OrderByComparator<AssetDisplayPageEntry> orderByComparator) {
+
+		return getPersistence().fetchByG_CN_Last(
+			groupId, classNameId, orderByComparator);
+	}
+
+	/**
+	 * Returns the asset display page entries before and after the current asset display page entry in the ordered set where groupId = &#63; and classNameId = &#63;.
+	 *
+	 * @param assetDisplayPageEntryId the primary key of the current asset display page entry
+	 * @param groupId the group ID
+	 * @param classNameId the class name ID
+	 * @param orderByComparator the comparator to order the set by (optionally <code>null</code>)
+	 * @return the previous, current, and next asset display page entry
+	 * @throws NoSuchDisplayPageEntryException if a asset display page entry with the primary key could not be found
+	 */
+	public static AssetDisplayPageEntry[] findByG_CN_PrevAndNext(
+			long assetDisplayPageEntryId, long groupId, long classNameId,
+			OrderByComparator<AssetDisplayPageEntry> orderByComparator)
+		throws com.liferay.asset.display.page.exception.
+			NoSuchDisplayPageEntryException {
+
+		return getPersistence().findByG_CN_PrevAndNext(
+			assetDisplayPageEntryId, groupId, classNameId, orderByComparator);
+	}
+
+	/**
+	 * Removes all the asset display page entries where groupId = &#63; and classNameId = &#63; from the database.
+	 *
+	 * @param groupId the group ID
+	 * @param classNameId the class name ID
+	 */
+	public static void removeByG_CN(long groupId, long classNameId) {
+		getPersistence().removeByG_CN(groupId, classNameId);
+	}
+
+	/**
+	 * Returns the number of asset display page entries where groupId = &#63; and classNameId = &#63;.
+	 *
+	 * @param groupId the group ID
+	 * @param classNameId the class name ID
+	 * @return the number of matching asset display page entries
+	 */
+	public static int countByG_CN(long groupId, long classNameId) {
+		return getPersistence().countByG_CN(groupId, classNameId);
+	}
+
+	/**
 	 * Returns the asset display page entry where groupId = &#63; and classNameId = &#63; and classPK = &#63; or throws a <code>NoSuchDisplayPageEntryException</code> if it could not be found.
 	 *
 	 * @param groupId the group ID