--- a/modules/apps/asset/asset-display-page-api/src/main/resources/com/liferay/asset/display/page/service/persistence/packageinfo
+++ b/modules/apps/asset/asset-display-page-api/src/main/resources/com/liferay/asset/display/page/service/persistence/packageinfo
@@ -1,1 +1,1 @@
-version 3.2.0
+version 3.3.0