--- a/modules/apps/asset/asset-display-page-service/src/main/resources/META-INF/sql/indexes.sql
+++ b/modules/apps/asset/asset-display-page-service/src/main/resources/META-INF/sql/indexes.sql
@@ -1,3 +1,4 @@
-create unique index IX_73E15F86 on AssetDisplayPageEntry (groupId, ctCollectionId, classNameId, classPK);
+create unique index IX_F3AB130A on AssetDisplayPageEntry (groupId, classNameId, ctCollectionId, classPK);
+create unique index IX_9920AB1F on AssetDisplayPageEntry (groupId, uuid_[$COLUMN_LENGTH:75$], ctCollectionId);
 create index IX_BFB8A913 on AssetDisplayPageEntry (layoutPageTemplateEntryId);
-create unique index IX_EF42C6CB on AssetDisplayPageEntry (uuid_[$COLUMN_LENGTH:75$], groupId, ctCollectionId);
+create index IX_DEA3F2DD on AssetDisplayPageEntry (uuid_[$COLUMN_LENGTH:75$]);