--- a/modules/apps/asset/asset-entry-rel-api/src/main/java/com/liferay/asset/entry/rel/service/AssetEntryAssetCategoryRelLocalService.java
+++ b/modules/apps/asset/asset-entry-rel-api/src/main/java/com/liferay/asset/entry/rel/service/AssetEntryAssetCategoryRelLocalService.java
@@ -6,6 +6,7 @@
 package com.liferay.asset.entry.rel.service;
 
 import com.liferay.asset.entry.rel.model.AssetEntryAssetCategoryRel;
+import com.liferay.asset.kernel.model.AssetEntry;
 import com.liferay.petra.function.UnsafeFunction;
 import com.liferay.petra.sql.dsl.query.DSLQuery;
 import com.liferay.portal.kernel.change.tracking.CTAware;
@@ -130,6 +131,9 @@ public interface AssetEntryAssetCategoryRelLocalService
 
 	public void deleteAssetEntryAssetCategoryRelByAssetCategoryId(
 		long assetCategoryId);
+
+	public void deleteAssetEntryAssetCategoryRelByAssetEntry(
+		AssetEntry assetEntry);
 
 	public void deleteAssetEntryAssetCategoryRelByAssetEntryId(
 		long assetEntryId);