--- a/modules/apps/asset/asset-entry-rel-api/src/main/java/com/liferay/asset/entry/rel/service/AssetEntryAssetCategoryRelLocalServiceUtil.java
+++ b/modules/apps/asset/asset-entry-rel-api/src/main/java/com/liferay/asset/entry/rel/service/AssetEntryAssetCategoryRelLocalServiceUtil.java
@@ -141,6 +141,12 @@ public class AssetEntryAssetCategoryRelLocalServiceUtil {
 			assetCategoryId);
 	}
 
+	public static void deleteAssetEntryAssetCategoryRelByAssetEntry(
+		com.liferay.asset.kernel.model.AssetEntry assetEntry) {
+
+		getService().deleteAssetEntryAssetCategoryRelByAssetEntry(assetEntry);
+	}
+
 	public static void deleteAssetEntryAssetCategoryRelByAssetEntryId(
 		long assetEntryId) {