--- a/modules/apps/asset/asset-entry-rel-api/src/main/resources/com/liferay/asset/entry/rel/service/packageinfo
+++ b/modules/apps/asset/asset-entry-rel-api/src/main/resources/com/liferay/asset/entry/rel/service/packageinfo
@@ -1,1 +1,1 @@
-version 2.0.0
+version 2.1.0