--- a/modules/apps/asset/asset-list-api/src/main/java/com/liferay/asset/list/service/AssetListEntryAssetEntryRelLocalService.java
+++ b/modules/apps/asset/asset-list-api/src/main/java/com/liferay/asset/list/service/AssetListEntryAssetEntryRelLocalService.java
@@ -142,6 +142,10 @@ public interface AssetListEntryAssetEntryRelLocalService
 			long assetListEntryId, long segmentsEntryId, int position)
 		throws PortalException;
 
+	public void deleteAssetListEntryAssetEntryRelByAssetEntryId(
+			long assetEntryId)
+		throws PortalException;
+
 	public void deleteAssetListEntryAssetEntryRelByAssetListEntryId(
 		long assetListEntryId);