… modules/apps/asset/asset-display-page-service/service.xml
… modules/apps/asset/asset-display-page-service/src/main/java/com/liferay/asset/display/page/service/persistence/impl/AssetDisplayPageEntryPersistenceImpl.java
… modules/apps/asset/asset-display-page-test/src/testIntegration/java/com/liferay/asset/display/page/service/persistence/test/AssetDisplayPageEntryPersistenceTest.java
… modules/apps/asset/asset-entry-rel-test/src/testIntegration/java/com/liferay/asset/entry/rel/test/AssetEntryAssetCategoryRelLocalServiceTest.java
… modules/apps/asset/asset-link-service/src/main/java/com/liferay/asset/link/service/impl/AssetLinkLocalServiceImpl.java
… modules/apps/asset/asset-list-api/src/main/java/com/liferay/asset/list/service/AssetListEntryAssetEntryRelLocalServiceWrapper.java
… modules/apps/asset/asset-list-api/src/main/resources/com/liferay/asset/list/service/packageinfo
… modules/apps/asset/asset-list-service/src/main/java/com/liferay/asset/list/internal/model/listener/AssetEntryModelListener.java
… modules/apps/asset/asset-list-service/src/main/java/com/liferay/asset/list/service/impl/AssetListEntryAssetEntryRelLocalServiceImpl.java
… modules/apps/asset/asset-service/src/main/java/com/liferay/asset/internal/service/AssetEntryLocalServiceWrapper.java
… modules/apps/layout/layout-api/src/main/java/com/liferay/layout/service/persistence/LayoutClassedModelUsagePersistence.java
… modules/apps/layout/layout-api/src/main/java/com/liferay/layout/service/persistence/LayoutClassedModelUsageUtil.java
… modules/apps/layout/layout-api/src/main/resources/com/liferay/layout/service/persistence/packageinfo
… modules/apps/layout/layout-service/service.xml
… modules/apps/layout/layout-service/src/main/java/com/liferay/layout/service/impl/LayoutClassedModelUsageLocalServiceImpl.java
… modules/apps/layout/layout-service/src/main/java/com/liferay/layout/service/persistence/impl/LayoutClassedModelUsagePersistenceImpl.java
… modules/apps/layout/layout-test/src/testIntegration/java/com/liferay/layout/service/persistence/test/LayoutClassedModelUsagePersistenceTest.java
… modules/apps/object/object-admin-rest-impl/src/main/java/com/liferay/object/admin/rest/internal/resource/v1_0/ObjectDefinitionResourceImpl.java
… modules/apps/object/object-api/src/main/java/com/liferay/object/definition/util/ObjectDefinitionThreadLocal.java
… modules/apps/object/object-api/src/main/java/com/liferay/object/tree/TreeFactory.java
… modules/apps/object/object-api/src/main/resources/com/liferay/object/definition/util/packageinfo
… modules/apps/object/object-api/src/main/resources/com/liferay/object/tree/packageinfo
… modules/apps/object/object-rest-test/src/testIntegration/java/com/liferay/object/rest/internal/manager/v1_0/test/DefaultObjectEntryManagerImplTest.java
… modules/apps/object/object-service/src/main/java/com/liferay/object/internal/deployer/ObjectDefinitionDeployerImpl.java
… modules/apps/object/object-service/src/main/java/com/liferay/object/internal/security/permission/resource/util/ObjectDefinitionResourcePermissionUtil.java
… modules/apps/object/object-service/src/main/java/com/liferay/object/internal/tree/TreeFactoryImpl.java
… modules/apps/object/object-service/src/main/java/com/liferay/object/model/impl/ObjectEntryImpl.java
… modules/apps/object/object-service/src/main/java/com/liferay/object/service/impl/ObjectDefinitionLocalServiceImpl.java
… modules/apps/object/object-service/src/main/java/com/liferay/object/service/impl/ObjectEntryLocalServiceImpl.java
… modules/apps/object/object-service/src/main/java/com/liferay/object/service/impl/ObjectEntryServiceImpl.java