This optimizes ObjectDefinition deletion performance down to linear.

Picture 1 shows the comparison for datasets that master baseline can finish within a waitable time (under 2 days).
![chart1](https://github.com/user-attachments/assets/6739c817-5d7f-4160-9a5c-58231970173e)

Comparing the largest dataset in this chart, which is 200K entries, the time dropped from 133595.332s (37.11hours) to 49.686s, **that is 99.96% off**, or in other term, **2688.79 times faster**. We did not run larger dataset for master baseline since it takes too much time.

Picture 2 shows PR only datapoints all the way to 2M dataset.
![chart2](https://github.com/user-attachments/assets/4cdcebd6-d99b-454b-829c-d58ed47a2fda)

With this PR, 2M enties deletion took 482.015sec = 8.03mins to finish.

By doing polynomial curve fitting, we get:

Tbaseline = 0.0014 * N^3 + 2.7513 * N^2 + 62.5485 * N − 158.0342

Tpr = 0.1884 * N + 3.9899

With this projection, if we were ever trying to run a 2M entries delete on master baseline, it would take:
Tb(2000) = 0.0014 * (2000)^3 + 2.7513 * (2000)^2 + 62.5485 * 2000 − 158.0342 = 22330138.9658s = 22330138.9658 / 60 / 60 / 24 = 258.45 days

Although it is mathematically pointless to compare a higher degree function with linear one, the ratio is always going to be infinite as the input is approaching infinite. But still to show the difference in real world, for a 2M entries deletion, this change makes it 22330138.9658/482.015 = 46326.65 times faster.

@victorhcunha 's detailed test results are at https://docs.google.com/spreadsheets/d/1igYs403Pp1cq3W_7Z_0p1vGZW-kIb1glvRmHk5HYLnE/edit?gid=0#gid=0


**For client db + client codebase test case, it took 36mins to finish**. This is worse than out 2M entries result for 2 reasons:

1) Client db has way more than 2M AssetEntry.
2) Client codebase is old that miss some other optimizations we have on master that noticeably improves non-db related AssetEntry removal performance.

I would suggest the client to upgrade to newer version of portal to gain the other performance improvements, rather than tracing down all missing pieces for hotfixes.

@gabrielwas @MarinhoFeliphe @JerryAdrianoo see the circular dependency issue fix at https://github.com/brianchandotcom/liferay-portal/commit/fc8ccd99bee5992cee6d183843e5d91fe614350e

CC @vicnate5 @wesleygong @marco-leo